requirements.txt
transformers
torch
gradio
pytorch-grad-cam
opencv-python-headless
pillow
numpy
matplotlib
scikit-image
scipy